Update TFM, libraries

+++ RaidBattlesBot/Configuration/BotConfiguration.cs
@@ -6,12 +6,12 @@ namespace RaidBattlesBot.Configuration
   public class BotConfiguration
   {
     public string[] BotTokens { get; set; }
-    public int? DefaultBotId { get; set; }
+    public long? DefaultBotId { get; set; }
     public TimeSpan Timeout { get; set; }
     public TimeSpan VoteTimeout { get; set; }
     public TimeSpan NotificationLeadTime { get; set; }
     public long? LogChatId{ get; set; }
-    public HashSet<int> BlackList { get; set; }
+    public HashSet<long> BlackList { get; set; }
     public string[] SkipDomains { get; set; }
   }
 }
