+++ RaidBattlesBot/Model/Vote.cs
@@ -5,9 +5,9 @@ namespace RaidBattlesBot.Model
 {
   public class Vote : ITrackable
   {
-    public int? BotId { get; set; }
+    public long? BotId { get; set; }
     public int PollId { get; set; }
-    public int UserId { get; set; }
+    public long UserId { get; set; }
     public string Username { get; set; }
     public string FirstName { get; set; }
     public string LastName { get; set; }
